All 12 handlers tested and passing against sandbox
Fixes applied during testing:
- VoucherHandler: row numbering (start from 1), amountGrossCurrency, GetStringField handles numbers, nested postings, locale-invariant parsing, snake_case field variants
- PaymentHandler: removed RawAmounts fallback, always use invoice amount from API
- TravelExpenseHandler: nested costs inside travel_expense entity, cost_items/costLines key variants, description->comments field fix
- InvoiceHandler: orderLines nested inside invoice entity, customer name from invoice.customer
- All handlers: int32->int64, CultureInfo.InvariantCulture for decimal parsing

New handlers: CreditNoteHandler, DeleteEntityHandler, ProjectHandler, TravelExpenseHandler, VoucherHandler
New scripts: Start-Agent.ps1, Test-Solve.ps1
Updated: copilot-instructions.md (dev workflow), README.md (testing docs)

+++ src/Handlers/CustomerHandler.cs
@@ -37,7 +37,7 @@ public class CustomerHandler : ITaskHandler
         _logger.LogInformation("Creating customer: {Name}", body.GetValueOrDefault("name"));
 
         var result = await api.PostAsync("/customer", body);
-        var customerId = result.GetProperty("value").GetProperty("id").GetInt32();
+        var customerId = result.GetProperty("value").GetProperty("id").GetInt64();
 
         _logger.LogInformation("Created customer ID: {Id}", customerId);
     }
@@ -49,7 +49,7 @@ public class CustomerHandler : ITaskHandler
         SetIfPresent(addr, cust, "postalCode");
         SetIfPresent(addr, cust, "city");
         if (cust.TryGetValue("countryId", out var countryId))
-            addr["country"] = new { id = int.Parse(countryId.ToString()!) };
+            addr["country"] = new { id = long.Parse(countryId.ToString()!) };
         return addr;
     }
